ASoC: cs40l25: Update handling for A2H delay/tuning controls
The A2H configuration can't be changed without informing the firmware
to reinitialise the algorithm and this command is only supported on
newer versions of the firmware. Writing these fields without following
this proceedure will at best have no effect and at worst cause the
algorithm to misbehave. Fix this issue, by not creating the offending
controls if an older version of the firmware is loaded and writing the
re-init command at the appropriate times.

Change-Id: I41a906a24d68a8ef19061bacf0f0068702e89d7a
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>