ASoC: cs40l25: Remove ASP routing from ASoC driver
Currently the ASoC driver attempts to setup the routing from the ASP
to the DSP, however, this is written once on probe and will thus be
lost whenever the chip enters hibernate. The primary driver currently
sets up half the routing from the ASP to the DSP, expand this to do
both channels and remove the now redundant routing from the ASoC
driver.

Change-Id: Iee36ed1669e6c63096a018f76154712c8242bade
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>