UPSTREAM: dt-bindings: soundwire: qcom: clarify data port bus parameters
Some of the parameters for data ports are not applicable or not implemented
in IP. So mark them as invalid/not applicable in DT so that controller is
aware of this.

Add comment to these bindings to provide more clarity on the values!

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Acked-by: Rob Herring <robh@kernel.org>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Link: https://lore.kernel.org/r/20210330144719.13284-2-srinivas.kandagatla@linaro.org
Signed-off-by: Vinod Koul <vkoul@kernel.org>
(cherry picked from commit 377785cc7c5d1dafadb1ae43c6d79ff934620f67)

BUG=b:192354788
TEST=build

Signed-off-by: Judy Hsiao <judyhsiao@chromium.org>
Change-Id: Ibd32495b3f4840d75615619322730d45cb3d7d5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2994912
Reviewed-by: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Judy Hsiao <judyhsiao@google.com>
Commit-Queue: Judy Hsiao <judyhsiao@google.com>