FROMGIT: media: venus: hfi,pm,firmware: Convert to block relative addressing
An upcoming silicon change places a number of existing blocks within the
Venus at different relative offsets to the base address of IO region.

In order to handle this difference this patch changes the address offsets
of the registers to function as offsets relative to the relevant sub-block
of registers within the IO region not the base address of the IO region.

As a result of this change venus_readl() and venus_writel() are deleted.

Co-developed-by: Dikshita Agarwal <dikshita@codeaurora.org>
Signed-off-by: Dikshita Agarwal <dikshita@codeaurora.org>
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Stanimir Varbanov <stanimir.varbanov@linaro.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit ff2a7013b3e6a3d34d2b5c7786b8a73093d25319
 https://git.linuxtv.org/media_tree.git master)

BUG=183962425
TEST=emerge-herobrine sys-kernel/chromeos-kernel-5_10

Change-Id: I84a8a34b2b01caebdd2f1de6487742583ca2980a
Signed-off-by: Vikash Garodia <vgarodia@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2825667
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Tested-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>