UPSTREAM: drm/tegra: sor: Do not leak runtime PM reference
It's theoretically possible for the runtime PM reference to leak if the
code fails anywhere between the pm_runtime_resume_and_get() and
pm_runtime_put() calls, so make sure to release the runtime PM reference
in that case.

Practically this will never happen because none of the functions will
fail on Tegra, but it's better for the code to be pedantic in case these
assumptions will ever become wrong.

Signed-off-by: Pavel Machek (CIP) <pavel@denx.de>
[treding@nvidia.com: add commit message]
Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 73a395c46704304b96bc5e2ee19be31124025c0c)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:187300590
TEST=Tested on volteer/zork/trogdor, build tested on the rest

Change-Id: I7d68e50828478d70a8c9be1a8edf322281c644a7