UPSTREAM: dt-bindings: net: qcom,ipa: add firmware-name property
Add a new optional firmware-name property to the IPA DT node.  It
is used only if the modem is not doing early initialization (i.e.,
if the modem-init property is not present).  Its value is the name
of the firmware file to use; if it's not specified, a default name
("ipa_fws.mdt") is used.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit d8604b209e9b3762280b8321162f0f64219d51c9
 https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git master)

BUG=b:175328613
TEST=Modem functional, suspend/resume work as expected

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: Ia87ed5888b9463d91152b8aba480909f6cc2b84f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2902562
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3154500