input: cs40l26: Improved gain control method
Using the AMP_VOL_PCM field to set global digital gain caused
degradation in the performance of the SVC algorithm.

Use the SOURCE_ATTENUATION control to set the digital gain
such that SVC is unaffected.

This control allows for the user to relay the digital gain
as a percentage to the driver (0 - 100) rather than
calculating the gain value on their own.

Change-Id: I06bf4173117582186992bcbd85aaea3a13f4d001
Signed-off-by: Fred Treven <fred.treven@cirrus.com>