CHROMIUM: Revert "UPSTREAM: drm/i915/gem: Drop relocation support on all new hardware (v6)"
This reverts commit ef3a6f12ebdb5f528453879d1859e87dd94abd7c.

This commit breaks media HW accel on Intel platforms since the current
20.4.5 Intel media driver uses relocation.

As mentioned in the reverted commit:
"This only leaves the media driver which is supposed to be switching to
softpin going forward."

The media driver commit adding softpin landed in main recently and
should be part of the the 21.2 release.
https://github.com/intel/media-driver/commit/144020c37770083974bedf59902b70b8f444c799
After upreving to 21.2, this revert can be dropped.

BUG=b:188899486
TEST=Check chrome://media-internals is using VaapiVideoDecoder for
playback
TOREVERT=b:189202986

Signed-off-by: Juston Li <juston.li@intel.com>
Change-Id: Iaeca62978bd2de9457ae971b559c8bef5d2f63f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2911364
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>