misc: cs40l25: Remove sleep before enabling classh
Remove a 50ms sleep before re-enabling classh after a clab
waveform. This delay is not needed because the interrupt
indicating that the waveform has completed is given after
the breaking of the waveform and not before. The reason this
sleep was put there was the assumption that we would have to
wait for the waveform to break before enabling classh.

Change-Id: I0f3ae2308f5b0fd999836344051265e10887ddf0
Signed-off-by: Paul Handrigan <Paul.Handrigan@cirrus.com>